Temporarily rename payload file
